requirements.txt
requests>=2.31.0
gitpython>=3.1.43
boto3>=1.34.3
pyyaml>=6.0
semver>=3.0.2
python-dotenv>=1.0.1
